#10860: Rename device conv files to align with rest of ops
